+++ operator/pkg/resources/target/statefulset_test.go
@@ -105,33 +105,43 @@ var _ = Describe("Statefulset Target Volumes and Mounts", Label("short"), func()
 	})
 	Describe("New Target with hostMount", func() {
 		It("should return no VolumeClaimTemplates but with volume mounts", func() {
-			specCopy := aisSpec.DeepCopy()
-			specCopy.Spec.StateStorageClass = nil
-			specCopy.Spec.HostpathPrefix = apc.Ptr("/node/data")
-			specCopy.Spec.TargetSpec.Mounts = []aisv1.Mount{
+			hostPathData := "/node/data"
+			testHostMount := []aisv1.Mount{
 				{
-					Path:        "/node/data",
+					Path:        hostPathData,
 					Size:        size,
 					UseHostPath: apc.Ptr(true),
 				},
 			}
+
+			specCopy := aisSpec.DeepCopy()
+
+			// Not used for this test, but explicitly set so we don't have a state volume claim to count
+			specCopy.Spec.StateStorageClass = nil
+			specCopy.Spec.HostpathPrefix = apc.Ptr("/etc/ais")
+
+			specCopy.Spec.TargetSpec.Mounts = testHostMount
 			result := NewTargetSS(specCopy, *specCopy.Spec.Size)
 			Expect(result).To(Not(BeNil()))
 			Expect(result.Spec.VolumeClaimTemplates).To(HaveLen(0))
-
 			Expect(result.Spec.Template.Spec.Volumes).To(HaveLen(7))
-			dataVolume := &v1.Volume{}
+
+			var dataVolume *v1.Volume
+			// Expect the host path data mount to be created with a fixed volume name
+			expectedName := "host-data-mount-0"
+			expectedHostPath := path.Join(hostPathData, specCopy.Namespace, specCopy.Name, apc.Target)
 			for _, dv := range result.Spec.Template.Spec.Volumes {
-				mnt := specCopy.Spec.TargetSpec.Mounts[0]
-				if dv.Name == mnt.GetMountName(specCopy.Name) {
+				if dv.Name == expectedName {
 					dataVolume = &dv
-					Expect(dataVolume.HostPath).To(Not(BeNil()))
-					Expect(dataVolume.HostPath.Path).To(Equal(path.Join(mnt.Path, specCopy.Namespace,
-						specCopy.Name, apc.Target)))
-					Expect(*dataVolume.HostPath.Type).To(Equal(v1.HostPathDirectoryOrCreate))
 				}
 			}
+			// Check that we found the expected data volume in the result with the right properties
 			Expect(dataVolume).To(Not(BeNil()))
+			Expect(dataVolume.Name).To(Equal(expectedName))
+			Expect(dataVolume.HostPath).To(Not(BeNil()))
+			Expect(dataVolume.HostPath.Path).To(Equal(expectedHostPath))
+			Expect(*dataVolume.HostPath.Type).To(Equal(v1.HostPathDirectoryOrCreate))
+
 			Expect(result.Spec.Template.Spec.Containers[0].VolumeMounts).To(HaveLen(5))
 		})
 	})